requirements.txt
python-dotenv
pandas
requests
discord.py
python-binance
